Fix build issues and complete MoveBar effects conversion

+++ src/WS2812FX/effects/PixelStackEffect.cpp
@@ -74,7 +74,7 @@ void PixelStackEffect::applyBackgroundFade(WS2812FX* strip) {
     // Apply speed-dependent fade to create smooth trails
     // Faster speeds get more fading to prevent blur
     const uint8_t MIN_FADE_AMOUNT = 2;
-    uint8_t fadeAmount = max(MIN_FADE_AMOUNT, effectSpeed >> 8);
+    uint8_t fadeAmount = max(MIN_FADE_AMOUNT, (uint8_t)(effectSpeed >> 8));
     
     fadeToBlackBy(strip->leds + runtime->start, runtime->length, fadeAmount);
 }
